New column widths for major edit/create forms.
After trial and error, the major edit and create forms are displayed
nicely with a column width of 35% for the labels and 65% for the input
fields. This means that most labels are on a single line and all
standard input elements do not exceed the table borders. The changes
are applied to all forms of the major objects (diseases, genes, etc.).

These figures are suitable for standard edit and create forms where
the table has a width of 760px with 14px reserved for the middle column
with help links, and when select boxes have the standard 75 characters
width for their content and text boxes have a width of 50 cols. This is
tested in Chromium and Firefox.