Do not throw secondary errors based on data type in object's checkfields().
The idea is that if checkfields() has reported an error earlier based
on the field's form field type, that error is probably more specific
(e.g. "checkbox value should be 0 or 1", "string field should be one of
these values") than the one reported for violating the data type (e.g.
"tinyint value should be smaller than 10").