Fix issue with display of link to reading frame checker.
Homepages for genes for which the refseq parser was run, but which have
no variants attached showed a distorted link to the reading frame
checker. The problem was that the HTML table for `Graphical displays
and utilities` was removed when no variants are present. This fix makes
sure that table is only removed when there is also no link to the
reading frame checker.