Created the lovd_getColumnMinMax() function.
Numeric columns are currently only checked for their format and the
length of the text value inside the field. The length is currently
checked just from the definition of the field. However, a TINYINT(3)
UNSIGNED doesn't allow for 999, even though the (3) suggests that it
does. This new function, lovd_getColumnMinMax() will calculate the
minimum and maximum values of a column and return these, so that values
can properly be checked.