Redefine FLOAT UNSIGNED into FLOAT.
MySQL 8 is now deprecating FLOAT UNSIGNED, together with DOUBLE UNSIGNED
and DECIMAL UNSIGNED. We have several FLOAT UNSIGNED columns in the
custom column definition array. Remove the UNSIGNED attribute and
updated the inc-upgrade.php script to remove this attribute in the
column definitions in the column table.

We don't need to actually alter the table; that's an expensive action
while not having much benefit - when the feature will be dropped, likely
MySQL will simply ignore the UNSIGNED flag.