When creating a gene by its symbol, recheck if we don't already have it.
When creating a gene by its HGNC ID, LOVD checks if we already have that
HGNC ID in the database. LOVD didn't do that if you use the symbol, it
only checked if we have that symbol already. So for genes who've changed
symbol, it was possible to open the create gene form using its new
symbol, after which you would get a query error. Now, LOVD checks if we
have the HGNC ID already once we fetched that from the HGNC.