Add option for additional permissions for LOVD.
As data sharing directly through the central LOVD APIs is more
complicated, it can conflict with the NC, ND, or SA licenses. Adding a
checkbox for allowing additional permissions to LOVD, it allows us to
relicense the data regardless of NC, ND, or SA selections for the rest
of the world. This will allow us to build a source of financial support.