Add dominant or recessive classification to the comments.
I first thought of including this information properly in VarioML's
InheritancePattern, but this requires a phenotype. This doesn't just
make coding it hard (phenotypes are stored at the Individual level and
they take their Inheritance already from the Disease(s) linked to the
Individual), it will also not always be available, e.g. for summary
records or classification records. So it's best to keep it light weight
and drop it in the comments, where also other parseable information
about clinical classifications are dropped.