Add reminder about inheritance.
I wanted to store Diseases having multiple forms of inheritance, but I
realized while implementing it that VarioML doesn't allow for an array
of inheritance values. In most cases, indeed there is only one value. So
decided to not implement anything; combined values are already stored as
a term just like values that are not recognized. Leaving a comment as a
reminder for this.