Remove spaces around select values when importing.
When importing, providing a value for a multiple selection list like
"PCR; SEQ" is valid, and " SEQ" gets stored. This is because
checkFields() trims the values, but the original values are left. The
value then mismatches, giving some minor issues. When the entry is
edited, the problem is also solved. But it's better to not import these
spaces.

This is now implemented into checkFields(), but the importer checks for
differences before checkFields() is run, so entries with spaces in them
will always trigger an update of the entry, even if there is nothing to
update. Because this is hard to fix, I'm leaving this.

Closes #539.