Relaxed the new XSS check a bit.
The less-than sign (<) wasn't allowed at all in fields anymore, but they
are in use for age fields and probably other phenotype fields as well.
So relaxed this check for all custom columns back to the previous XSS
check. The newer check for no less-than at all, remains for non-custom
columns.