Update getVariantInfo to use 1 regex instead of 2
To make getVariantInfo simpler and easier to use, the two regular
expressions have been pulled into one. Secondly, the priority has
been put on finding the positions, so that variants will be sorted
accordingly even if the variant as a whole does not follow HGVS
syntax and is therefore ambiguous or implausable. Thirdly, more
warning messages are being returned to help the user understand
where the problems in their variant description lie.