Fix and expand the new getVariantInfo()
This new commit fixes the bugs in getVariantInfo and adds multiple
functionalities. The function can now deal with repeats, and
recognises when suffixes are rightly given to variants of a type
other than ins or delins. Most of all, this commit will allow
getVariantInfo() to give truly meaningful warning and error
messages when poorly formatted variants are given to it.

Secondly, all sorts of cases have been added to the getVariantInfo
testcase, to make sure even the strangest of formats are being
tested.