Removed broken check for a situation that is too rare.
The code contained a check aimed at variants starting with a
parenthesis, while the parentheses in the variant are also unbalanced.
This type of variant, "(c.123del", doesn't actually currently occur in
the database, and the fix is broken and actually malforms other
variants, like "(c.123del)". Mixing logical AND with OR without the use
of parantheses creates confusing results.