Add a more in depth handling of reference sequences
This commit allows getVariantInfo() to make sure the right reference
sequence is given to match a variant's DNA type. It makes sure even
more reference types are accepted (now adding Ensemble and LRG
sequences). The function now returns errors if any mistakes were made.
Also, new checks were added to the test case to make sure the new code
works as it should.

Note: as it stands, variant with a wrongly formatted reference sequences
cannot be interpreted by getVariantInfo() as it stands. In the future
this might be handled better by adding a functionality which seperates
'variant' from reference sequence by the ':'; and checking the variant
even if the reference sequence does not match expectations.