Check for wrongly formatted reference sequences
GetVariantInfo() will now return an error if it has found something
that seems to be a reference sequence, but does not match the
expected format. It does so by a very broad first regular expression,
followed by a second one which dives into the details.

Adding this extra check not only makes sure users are being warned if
their reference sequence is not correct, but it also means that now
the variant AFTER the reference is being put through to the rest of
the checks correctly even if the first half was not recognised.