Remove $sReference as a function argument.
Doubting the implementation, I tested several variants and found
that they caused infinite recursion (e.g., "NC_123456.1:(g.123A=)").
Decided that it was easier to just remove the argument and redo the
parsing of the reference sequence each time the function is called. The
feature's speed improvement was marginal compared to the rest of the
function, but anyway I simplified the regexp as well to reduce parsing
time.

Also, found one lovd_fixHGVS() call that didn't prefix the call with the
reference sequence.