Accept the suffix insN[1..] instead of ins(1..)
Note: although this update will allow variants using N[20] or similar
approaches, it does not yet spot mistakes such as insertions as
follows: c.1_2insNNNNNN (too many Ns; should be rewritten as
c.1_2insN[6]). These variants will be passed on as HGVS=true. Adding a
warning in this cases should be done when there is a clear idea of
what the cut-off length is to decide between the two formats.