Change EMISSINGREFERENCE to EWRONGREFERENCE
After discussion it has been decided that EWRONGREFERENCE is a more
meaningful name for this error than EMISSINGREFERENCE was. Our next
step will be to also apply this error in cases of variants which
hold intronic positions in their insertions (i.e. the suffixes of
variants of type (del)ins), similarly to how we are already applying
it to the main positions.