Improved handling of WSUFFIXFORMAT.
Two changes;
- There's no need to check the suffix format, when a suffix isn't even
allowed.
- I wanted to improve the "inserted/affected sequence" part of the
warning to implement a choice of one of the two, but I realized often
the suffix didn't even contain a sequence. This may be confusing,
so I decided to reformat the warning message and to indicate what part
of the string is not according to the guidelines.