Fixed recognition of met= variants and added tests for related variants.
Met= variants were recognized as '=' type, so no change. Technically
this is correct, but none of the methylation-related variants change the
sequence. It is therefore more correct to have met= recognized as type
'met'. Moved the code around a bit to achieve this, and added tests for
met=, lom, and gom variants.