Improved check for missing pipe, renamed to an error, and added a fix.
Check was just based on finding the variant type anywhere within the
variant. Now matching more carefully, also demanding a position just
before the variant type. Renamed the warning to an error as there are no
positions returned and there was no guarantee that the variant was
fixable ("asdfgom" also matched). lovd_fixHGVS() wasn't actually
handling these variants yet, so added a fix for it, and a test.