Fixed four issues handling questionmarks by rebuilding the whole code.
The code contained an if() that wasn't reached, another if() with a
typo, and some other errors that I didn't locate yet. The following
issues have been solved:

g.(5_?)_?del
Please rewrite the positions (5_?)_? to (5_0).

g.(?_?)_(10_?)del
Please rewrite the positions (?_?)_(10_?) to ?_10.

g.(5_?)_(?_?)del
Please rewrite the positions (?_?)_(?_?) to ?.

g.(?_5)_(?_?)del
Please rewrite the positions (?_5)_(?_?) to ?_?.

The code now works not on checking for all possibilities, but by
breaking down all possibilities in smaller pieces, and handling them one
by one. All issues can be solved in five simple steps.