Fixed bug in questionmark handling, rebuilt the code.
lovd_fixHGVS() was mishandling two variants from the tests;

g.(?_10)del       was corrected to   g.?_10)del
c.(5-1_10+1)del   was corrected to   c.(5_10+1)del

Instead of trying to fix the error by fixing the code, decided to just
let lovd_getVariantInfo() do all the work. They already fixed the
positions, so why should we try to figure out the problem again? Just
take their suggestion and apply it.

Because the removal of questionmarks was coupled with code sorting the
positions, it's quite a diff. The code is now simpler, quicker, and with
less nesting, causing an indentation change that make the diff very
large. The position sorting code has not been touched.