Clarify when a suffix is never allowed vs not understood.
lovd_getVariantInfo() didn't clarify when a suffix was not
parseable, like "delXX", or never allowed, like, "A>C--". Both
threw a WSUFFIXGIVEN. That makes it harder for lovd_fixHGVS() to
see the difference. Clarify this; WSUFFIXGIVEN means suffix is
never allowed for this type (i.e., substitution), or allowed but
redundant. WSUFFIXFORMAT means suffix might be allowed for this
type, but not parseable.